Moving plugin logging loader out of bg_utils package
